New version of batoco
The biggest outward change is the new switch -m which allows a target machine to be selected. Targets are ZX80,ZX81,Lambda,Spectrum,Timex and Plus3. This is still a work in progress.

Spectrum support is pretty much there.
ZX80 and ZX81 have only received basic testing, bugs will exist.
Timex support is pretty much the same as the Spectrum with just a few added keywords.
Lambda support is currently broken, but 95% there.
Plus3 support is work in progress.

Initial code for a basic PreProcessor to allow one code source to support different machines.
Spanish characters ¡,Ñ,¿,€(instead of old Spanish currency sign) and ñ.