Fix for broken Spectrum builds
Fixed dodgy logic in frexp function which incorrectly tried to process the zero of LET a=0 as a floating point instead of an Integer for Spectrum builds. Spectrum builds should work correctly again.

Although there is a deal of code added to the preprocessor, it currently is not working. This is only a bug fix release not a new feature release.